+++ NEWS.md
@@ -46,6 +46,8 @@
   
 * Time-Dependent ROC AUC estimation for right-censored data can now be 
   calculated with `roc_auc_survival()`.
+  
+* `metric_set()` can now be used with a combination of dynamic and static survival metrics.
 
 # yardstick 1.1.0